Moved question files to hard-q9
